Take code out from visual_lab.py into dedicated widgets modules
